fix(wazuh-ai-assistant): rename search_alerts_by_rule_id to search_alerts_by_rule_title
wazuh.rule.id is a UUID in 5.0, not the short numeric id an analyst
could name in a question (e.g. "5710"), so a "search by rule ID" tool
premise no longer holds. Pivoted the tool to match on wazuh.rule.title
instead, which stays a human-readable, reachable identifier.
