+++ docs/ref/modules/ai-assistant/tool-catalog.md
@@ -27,17 +27,17 @@ Key properties:
 
 ## The 29 tools
 
-| Category                 | Tools                                                                                                                                                                                                                                                                                |
-| ------------------------ | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------ |
-| Agents                   | `get_active_agents`, `get_disconnected_agents`                                                                                                                                                                                                                                       |
-| Alerts                   | `get_critical_alerts`, `get_alerts_by_time`, `get_top_rules`, `get_security_summary`, `get_brute_force`, `get_suspicious_powershell`, `search_alerts_by_agent`, `search_alerts_by_multiple_agents`, `search_alerts_by_os`, `search_alerts_by_rule_group`, `search_alerts_by_rule_id` |
-| Vulnerabilities          | `get_vulnerabilities`, `get_critical_vulnerabilities`, `get_vulnerabilities_by_agent`, `get_vulnerability_by_cve`                                                                                                                                                                    |
-| FIM                      | `get_fim_files`                                                                                                                                                                                                                                                                      |
-| SCA                      | `get_sca_results`, `get_sca_checks`                                                                                                                                                                                                                                                  |
-| MITRE ATT&CK             | `get_mitre_alerts`, `get_mitre_summary`                                                                                                                                                                                                                                              |
-| Inventory (syscollector) | `get_agent_os`, `get_agent_packages`, `get_agent_ports`, `get_agent_processes`                                                                                                                                                                                                       |
-| Compliance               | `get_pci_dss_alerts`, `get_pci_dss_summary`                                                                                                                                                                                                                                          |
-| Free search              | `search_wazuh_data` (the escape hatch)                                                                                                                                                                                                                                               |
+| Category                 | Tools                                                                                                                                                                                                                                                                                   |
+| ------------------------ | --------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
+| Agents                   | `get_active_agents`, `get_disconnected_agents`                                                                                                                                                                                                                                          |
+| Alerts                   | `get_critical_alerts`, `get_alerts_by_time`, `get_top_rules`, `get_security_summary`, `get_brute_force`, `get_suspicious_powershell`, `search_alerts_by_agent`, `search_alerts_by_multiple_agents`, `search_alerts_by_os`, `search_alerts_by_rule_group`, `search_alerts_by_rule_title` |
+| Vulnerabilities          | `get_vulnerabilities`, `get_critical_vulnerabilities`, `get_vulnerabilities_by_agent`, `get_vulnerability_by_cve`                                                                                                                                                                       |
+| FIM                      | `get_fim_files`                                                                                                                                                                                                                                                                         |
+| SCA                      | `get_sca_results`, `get_sca_checks`                                                                                                                                                                                                                                                     |
+| MITRE ATT&CK             | `get_mitre_alerts`, `get_mitre_summary`                                                                                                                                                                                                                                                 |
+| Inventory (syscollector) | `get_agent_os`, `get_agent_packages`, `get_agent_ports`, `get_agent_processes`                                                                                                                                                                                                          |
+| Compliance               | `get_pci_dss_alerts`, `get_pci_dss_summary`                                                                                                                                                                                                                                             |
+| Free search              | `search_wazuh_data` (the escape hatch)                                                                                                                                                                                                                                                  |
 
 On Wazuh 5.0 the tools read from the 5.0 data layer: the `wazuh-events-v5-*` event indices,
 `wazuh-findings-v5-*`, and the `wazuh-states-*` state indices (vulnerabilities, FIM, SCA,
